Midiio

+++ src/main/kotlin/de/pans/dot2/Dot2Mapper.kt
@@ -2,10 +2,10 @@ package de.pans.dot2
 
 object Dot2Mapper {
 
-    fun map(bytes: List<Byte>): List<Byte> {
+    fun map(bytes: List<Int>): List<Int> {
         val bytes = bytes.toMutableList()
         if (MappingSettings.isBound(bytes[1])) {
-            bytes[0] = 144.toByte()
+            bytes[0] = 144
             bytes[1] = MappingSettings.getBind(bytes[1])
 
             return bytes
